# SIMD: Optimize the performance of einsum's submodule dot .