## Introduction

This is the fourth part of #17049, There has no impact on SSE2 platform and about **_10%~64%_** increased performance in ARM, the performance increased by **_27%_** with avx2 instrument . Here is the benchmark results:

## Benchmark

Here is the ASV benchmark result.
<details>
<summary>SSE2 enabled</summary>
<pre><code>
· Creating environments
· Discovering benchmarks
·· Uninstalling from virtualenv-py3.7-Cython
·· Building 350670fd <einsum-dot> for virtualenv-py3.7-Cython
·· Installing 350670fd <einsum-dot> into virtualenv-py3.7-Cython
· Running 28 total benchmarks (2 commits * 1 environments * 14 benchmarks)
[  0.00%] · For numpy commit 9e26d1d2 <master> (round 1/2):
[  0.00%] ·· Building for virtualenv-py3.7-Cython
[  0.00%] ·· Benchmarking virtualenv-py3.7-Cython
[  1.79%] ··· Running (bench_linalg.Einsum.time_einsum_contig_contig--)..............
[ 25.00%] · For numpy commit 350670fd <einsum-dot> (round 1/2):
[ 25.00%] ·· Building for virtualenv-py3.7-Cython
[ 25.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 26.79%] ··· Running (bench_linalg.Einsum.time_einsum_contig_contig--)..............
[ 50.00%] · For numpy commit 350670fd <einsum-dot> (round 2/2):
[ 50.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 51.79%] ··· bench_linalg.Einsum.time_einsum_contig_contig                                                                                                                       ok
[ 51.79%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   146±8μs
               numpy.float64   247±10μs
              =============== ==========

[ 53.57%] ··· bench_linalg.Einsum.time_einsum_contig_outstride0                                                                                                                   ok
[ 53.57%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   256±30μs
               numpy.float64   469±20μs
              =============== ==========

[ 55.36%] ··· bench_linalg.Einsum.time_einsum_mul                                                                                                                                 ok
[ 55.36%] ··· =============== =============
                   dtype
              --------------- -------------
               numpy.float32   1.41±0.04ms
               numpy.float64    2.79±0.1ms
              =============== =============

[ 57.14%] ··· bench_linalg.Einsum.time_einsum_multiply                                                                                                                            ok
[ 57.14%] ··· =============== ===========
                   dtype
              --------------- -----------
               numpy.float32   67.4±10μs
               numpy.float64    81.8±1μs
              =============== ===========

[ 58.93%] ··· bench_linalg.Einsum.time_einsum_noncon_contig_contig                                                                                                                ok
[ 58.93%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   46.7±2μs
               numpy.float64   46.4±1μs
              =============== ==========

[ 60.71%] ··· bench_linalg.Einsum.time_einsum_noncon_contig_outstride0                                                                                                            ok
[ 60.71%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   35.8±1μs
               numpy.float64   35.9±2μs
              =============== ==========

[ 62.50%] ··· bench_linalg.Einsum.time_einsum_noncon_mul                                                                                                                          ok
[ 62.50%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32    41.6±2μs
               numpy.float64   42.8±0.9μs
              =============== ============

[ 64.29%] ··· bench_linalg.Einsum.time_einsum_noncon_multiply                                                                                                                     ok
[ 64.29%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   70.4±3μs
               numpy.float64   78.6±2μs
              =============== ==========

[ 66.07%] ··· bench_linalg.Einsum.time_einsum_noncon_outer                                                                                                                        ok
[ 66.07%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32   11.2±0.5ms
               numpy.float64   22.6±0.2ms
              =============== ============

[ 67.86%] ··· bench_linalg.Einsum.time_einsum_noncon_sum_mul                                                                                                                      ok
[ 67.86%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   67.8±2μs
               numpy.float64   76.6±2μs
              =============== ==========

[ 69.64%] ··· bench_linalg.Einsum.time_einsum_noncon_sum_mul2                                                                                                                     ok
[ 69.64%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   71.3±5μs
               numpy.float64   77.3±1μs
              =============== ==========

[ 71.43%] ··· bench_linalg.Einsum.time_einsum_outer                                                                                                                               ok
[ 71.43%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   25.9±1ms
               numpy.float64   51.2±1ms
              =============== ==========

[ 73.21%] ··· bench_linalg.Einsum.time_einsum_sum_mul                                                                                                                             ok
[ 73.21%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   62.5±1μs
               numpy.float64   68.7±3μs
              =============== ==========

[ 75.00%] ··· bench_linalg.Einsum.time_einsum_sum_mul2                                                                                                                            ok
[ 75.00%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   64.8±2μs
               numpy.float64   67.2±2μs
              =============== ==========

[ 75.00%] · For numpy commit 9e26d1d2 <master> (round 2/2):
[ 75.00%] ·· Building for virtualenv-py3.7-Cython
[ 75.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 76.79%] ··· bench_linalg.Einsum.time_einsum_contig_contig                                                                                                                       ok
[ 76.79%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   160±4μs
               numpy.float64   266±20μs
              =============== ==========

[ 78.57%] ··· bench_linalg.Einsum.time_einsum_contig_outstride0                                                                                                                   ok
[ 78.57%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   266±9μs
               numpy.float64   466±20μs
              =============== ==========

[ 80.36%] ··· bench_linalg.Einsum.time_einsum_mul                                                                                                                                 ok
[ 80.36%] ··· =============== =============
                   dtype
              --------------- -------------
               numpy.float32   1.35±0.04ms
               numpy.float64   2.80±0.07ms
              =============== =============

[ 82.14%] ··· bench_linalg.Einsum.time_einsum_multiply                                                                                                                            ok
[ 82.14%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   73.0±5μs
               numpy.float64   83.1±1μs
              =============== ==========

[ 83.93%] ··· bench_linalg.Einsum.time_einsum_noncon_contig_contig                                                                                                                ok
[ 83.93%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   48.2±3μs
               numpy.float64   47.4±5μs
              =============== ==========

[ 85.71%] ··· bench_linalg.Einsum.time_einsum_noncon_contig_outstride0                                                                                                            ok
[ 85.71%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32   35.1±0.7μs
               numpy.float64    37.6±1μs
              =============== ============

[ 87.50%] ··· bench_linalg.Einsum.time_einsum_noncon_mul                                                                                                                          ok
[ 87.50%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   41.2±1μs
               numpy.float64   42.4±2μs
              =============== ==========

[ 89.29%] ··· bench_linalg.Einsum.time_einsum_noncon_multiply                                                                                                                     ok
[ 89.29%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   69.3±1μs
               numpy.float64   84.7±8μs
              =============== ==========

[ 91.07%] ··· bench_linalg.Einsum.time_einsum_noncon_outer                                                                                                                        ok
[ 91.07%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32   11.4±0.1ms
               numpy.float64   22.9±0.4ms
              =============== ============

[ 92.86%] ··· bench_linalg.Einsum.time_einsum_noncon_sum_mul                                                                                                                      ok
[ 92.86%] ··· =============== ===========
                   dtype
              --------------- -----------
               numpy.float32    70.5±3μs
               numpy.float64   78.7±10μs
              =============== ===========

[ 94.64%] ··· bench_linalg.Einsum.time_einsum_noncon_sum_mul2                                                                                                                     ok
[ 94.64%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   70.5±4μs
               numpy.float64   81.8±5μs
              =============== ==========

[ 96.43%] ··· bench_linalg.Einsum.time_einsum_outer                                                                                                                               ok
[ 96.43%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32   26.5±0.9ms
               numpy.float64    50.9±1ms
              =============== ============

[ 98.21%] ··· bench_linalg.Einsum.time_einsum_sum_mul                                                                                                                             ok
[ 98.21%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   60.9±2μs
               numpy.float64   68.1±3μs
              =============== ==========

[100.00%] ··· bench_linalg.Einsum.time_einsum_sum_mul2                                                                                                                            ok
[100.00%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   63.5±3μs
               numpy.float64   65.5±2μs
              =============== ==========


BENCHMARKS NOT SIGNIFICANTLY CHANGED.
</code></pre>
</details>

<details>
<summary>AV2 enabled</summary>
<pre><code>
· Creating environments
· Discovering benchmarks
·· Uninstalling from virtualenv-py3.7-Cython
·· Installing c32f60e3 <einsum-dot> into virtualenv-py3.7-Cython
· Running 28 total benchmarks (2 commits * 1 environments * 14 benchmarks)
[  0.00%] · For numpy commit 9e26d1d2 <master> (round 1/2):
[  0.00%] ·· Building for virtualenv-py3.7-Cython
[  0.00%] ·· Benchmarking virtualenv-py3.7-Cython
[  1.79%] ··· Running (bench_linalg.Einsum.time_einsum_contig_contig--)..............
[ 25.00%] · For numpy commit c32f60e3 <einsum-dot> (round 1/2):
[ 25.00%] ·· Building for virtualenv-py3.7-Cython
[ 25.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 26.79%] ··· Running (bench_linalg.Einsum.time_einsum_contig_contig--)..............
[ 50.00%] · For numpy commit c32f60e3 <einsum-dot> (round 2/2):
[ 50.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 51.79%] ··· bench_linalg.Einsum.time_einsum_contig_contig                                                                                                                       ok
[ 51.79%] ··· =============== =========
                   dtype
              --------------- ---------
               numpy.float32   136±6μs
               numpy.float64   191±6μs
              =============== =========

[ 53.57%] ··· bench_linalg.Einsum.time_einsum_contig_outstride0                                                                                                                   ok
[ 53.57%] ··· =============== =========
                   dtype
              --------------- ---------
               numpy.float32   252±7μs
               numpy.float64   446±8μs
              =============== =========

[ 55.36%] ··· bench_linalg.Einsum.time_einsum_mul                                                                                                                                 ok
[ 55.36%] ··· =============== =============
                   dtype
              --------------- -------------
               numpy.float32   1.41±0.05ms
               numpy.float64   2.95±0.04ms
              =============== =============

[ 57.14%] ··· bench_linalg.Einsum.time_einsum_multiply                                                                                                                            ok
[ 57.14%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   86.9±4μs
               numpy.float64   96.2±2μs
              =============== ==========

[ 58.93%] ··· bench_linalg.Einsum.time_einsum_noncon_contig_contig                                                                                                                ok
[ 58.93%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   49.8±2μs
               numpy.float64   58.6±4μs
              =============== ==========

[ 60.71%] ··· bench_linalg.Einsum.time_einsum_noncon_contig_outstride0                                                                                                            ok
[ 60.71%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32    37.4±2μs
               numpy.float64   36.8±0.8μs
              =============== ============

[ 62.50%] ··· bench_linalg.Einsum.time_einsum_noncon_mul                                                                                                                          ok
[ 62.50%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   43.2±2μs
               numpy.float64   43.6±6μs
              =============== ==========

[ 64.29%] ··· bench_linalg.Einsum.time_einsum_noncon_multiply                                                                                                                     ok
[ 64.29%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   86.5±5μs
               numpy.float64   101±5μs
              =============== ==========

[ 66.07%] ··· bench_linalg.Einsum.time_einsum_noncon_outer                                                                                                                        ok
[ 66.07%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32   11.4±0.3ms
               numpy.float64   23.0±0.2ms
              =============== ============

[ 67.86%] ··· bench_linalg.Einsum.time_einsum_noncon_sum_mul                                                                                                                      ok
[ 67.86%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   68.0±2μs
               numpy.float64   75.6±3μs
              =============== ==========

[ 69.64%] ··· bench_linalg.Einsum.time_einsum_noncon_sum_mul2                                                                                                                     ok
[ 69.64%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   69.9±2μs
               numpy.float64   77.5±2μs
              =============== ==========

[ 71.43%] ··· bench_linalg.Einsum.time_einsum_outer                                                                                                                               ok
[ 71.43%] ··· =============== ============
                   dtype
              --------------- ------------
               numpy.float32   25.8±0.7ms
               numpy.float64   53.1±0.4ms
              =============== ============

[ 73.21%] ··· bench_linalg.Einsum.time_einsum_sum_mul                                                                                                                             ok
[ 73.21%] ··· =============== ==========
                   dtype
              --------------- ----------
               numpy.float32   62.5±1μs
               numpy.float64   67.0±2μs
              =============== ==========

[ 75.00%] ··· bench_linalg.Einsum.time_einsum_sum_mul2                                                                                                                            ok
[ 75.00%] ··· =============== ==========
                   dtype
              --------------- ----------
…