@mattip 2020-12-14
Nice, removed ~100 lines of code and increased performance. @seiko2plus could you take a look?